Working NES controller
