+++ DE1_SoC_demo/RTL/quartus/DE1_SoC_demo.qsf
